+++ docs/changelog.rst
@@ -3,9 +3,24 @@ Change log
 
 .. currentmodule:: picozero
 
-0.5.1 - 2025-11-24
+0.6.0 - 2025-11-26
+-----------
 
-+ Fix to incorrect example in documentation.
++ Introduced ``Stepper`` class for stepper motors
++ Introduced ``TouchSensor`` class for capacitive touch sensors
++ Updated tests and documentation
++ Other minor bug fixes
+
+
+0.5.2 - 2025-11-26
+-----------
+
++ Fixed 404 in manual install instructions
+
+0.5.1 - 2025-11-24
+-----------
+
++ Fix to incorrect example in documentation
 
 0.5.0 - 2025-10-31
 -----------